Implement generator delegation
Summary:
Implements https://github.com/facebook/hhvm/issues/6040

Generator delegation adds support for the `yield from` syntax, where the outer generator "delegates" the yielding of keys/values to its subgenerator. At a high level, you can think of `yield from gen1()` to be equivalent to `foreach(gen1() as $key => $val) { yield $key => $val; }`. To make this happen, I added 3 new bytecodes - `ContAssignDelegate`, `ContEnterDelegate`, and `YieldFromDelegate`. The syntax `yield from <expr>` then evaluates to

```
ContAssignDelegate <expr>
ContEnterDelegate
YieldFromDelegate
ContAssignDelegate NULL
```

The job of ContAssignDelegate is exactly what you think. ContEnterDelegate is in charge of continuing the delegate generator's execution (its basically a shim over ContEnter with an extra branch to see if `<expr>` is an interator). YieldFromDelegate is where most of the interesting logic happens. For iterators it basically functions as a normal yield, but for generators we can't copy any data from the delegate. This is because the delegate generator could be advanced outside of the context of delegating generator, in which case the data would get out of sync. This is why I had to special case the `ContCurrent` and `ContKey` bytecodes as well - they need to understand that if we have a delegate currently assigned then we need to get our values from there instead. Since YieldFromDelegate doesnt need to actually store any yielded values for generators, all it does is clean up the generator's tracked state.

For ease of implementation, all of the `yield from` related bytecodes are only implemented in the interpreter.

Reviewed By: paulbiss

Differential Revision: D2681385

fb-gh-sync-id: 5eb9e5069d8b9b3c4aa2ef89fc0eeb40563ac603