Move the zend-imported crypto files into C++
Summary:
I want to fix some locking here as part of re-porting this to OS X, and `Mutex` is the way to do that, which means this needs to be C++. It's better for consistency IMO anyways, there's a bit of nasty `extern` cleanup that can happen here.

Apologies to Orvid if this breaks `php-crypt_r.cpp`; I don't have a way to test that file right now. The others should be fine though.

Reviewed By: paulbiss

Differential Revision: D2669838

fb-gh-sync-id: d3037ecb4713dc0c9bff202b58bd1de1c4b13521