Support `crypt()` compatibility layer on OS X
Summary:
Orvid already pulled in code from PHP to support the glibc extended features of `crypt()` on MSVC. PHP supports that layer on other OSes too, but in doing so, Orvid specalized it for Windows since that was the one platform we cared about at the time. It turns out that OS X lacks the same glibc extended features, and so needs this layer too. This diff does a little bit of cleanup and un-specalization of Orvid's work. The only completely new batch of code here is copy-pasted directly from PHP's implementation (though some of the other changes here are very heavily based on what PHP does).

This probably breaks MSVC in some stupid but minor way -- sorry in advance Orvid, I have no way of testing Windows right now.

Reviewed By: paulbiss

Differential Revision: D2672015

fb-gh-sync-id: 3983b2acd5d2eb1b8d9591778a6c11eb2f3ed077