Respect request timeout when SleepWaitHandles are used
Summary:
The current implementation bounds the I/O wait time by either the
earliest finishing SleepWaitHandle or remaining request time. If
there is a SleepWaitHandle with excessive sleep duration, the request
timeout would not be respected [*].

Let's fix this and respect the request timeout.

[*] Actually, it "works" as expected due to a bug in glibc causing
std::this_thread::sleep_until() to terminate on signal, combined
with a fact that the timeout mechanism is signal based. The glibc
bug is going to be fixed though:

https://gcc.gnu.org/ml/libstdc++/2015-11/msg00049.html

Reviewed By: interwq

Differential Revision: D2740442

fb-gh-sync-id: 61e59593a214c3bf90ade59ddc57c1f1da138263