Backout "Get watchman updates via subscription rather than polling"
Summary:
We don't seem to be getting all the updates we should. I suspect this
may be triggered when the same file is changed multiple times in quick
succession; we seem to be getting only one update even if the file is changed
again shortly after that update.

Backing out for now to unblock the deploy.

Reviewed By: dlreeves

Differential Revision: D2742519

fb-gh-sync-id: 3dac6bf3c44fa7ea0f5b5ea756f8498be90a6bea