Export symbols more selectively
Summary:
hhvm is currently linked with --export-dynamic, which means that all
external symbols are put in the dynamic symbol table. This is
wasteful, slows down startup, and should also prevent the code from
being stripped via --gc-sections. Due to a bug in older versions of
binutils, gold stripped them anyway.

The bug has been fixed in 2.25, which led to a huge increase in the
binary size, and eventually to the discovery that we were exporting
everything. So stop exporting everything, and just export the (small
number of) symbols we actually need. This actually saves over 10M on
the binary size too.

Reviewed By: maksfb

Differential Revision: D2747334

fb-gh-sync-id: 0088eca1ccb31c69070d3a9bb43c0454878a363f