Rename use_watchman_2 to use_watchman
Summary:
I originally had to switch to use_watchman_2 because I had suspected that there
were bugs in the original hh_server code that used watchman. However, we've
since found out that the bug is likely to be in watchman, and regardless the
affected hh_server version is old enough that it doesn't matter.

Reviewed By: dabek

Differential Revision: D2747798

fb-gh-sync-id: 156bf153cc67cb3545ce459e94047a0566fcd219