Cleanup some and/test patterns
Summary:
After the FCallAwait diff, eager returns from async functions generate
a guard that looks like:

  loadzql mem, reg
  andq 0x80000000, reg, dst
  testq dst, dst, sf
  jnz sf

This resulted in x64 code like:

  mov eax, dword ptr [rbp+0x1c]
  mov ecx, 0x80000000
  and rax, rcx
  test rax, rax
  jz 0x176003f9

This adds a simplifier that collapses the andq/testq into a single
testq (when dst is not used again), changes the immediate to be signed
(to make constant folding possible) and adds the missing fold-imm for
testq so that we end up with:

  mov eax, dword ptr [rbp+0x1c]
  test rax, 0xffffffff80000000
  jz 0x16c016b8

This could still be better: because we know the high bits of rax are
zero, we could use a 32 bit test (saving a rex byte). Also, we could
combine the load and the test; or for this particular constant, we could
do "test eax, eax" and check for a negative result.

I'll do some of this in a followup.

Reviewed By: swtaarrs

Differential Revision: D2734278

fb-gh-sync-id: 20c3711bb1a6b4d0d8b66c3dfe55d177b2357527