clean up some array functions
Summary:
Historically PackedArray was implemented as a MixedArray, and many helpers lived in the MixedArray class, e.g., MixedArray::MakeUncountedPacked().  This diff tries to move them to the PackedArray class, and also extract some helpers common to all array classes, e.g., ConvertTvToUncounted().

This is intended to be a mechanical change that doesn't affect behavior.  It is useful when we add a new array kind (will do in another diff).

Reviewed By: edwinsmith

Differential Revision: D2741773

fb-gh-sync-id: be0684933480fe2db469084537c079b2e9f9c0f8