Fix SOCKET_E* constants
Summary:
The HHVM_RC_INT macro assumes the constant name in the raw,
but REGISTER_LONG_CONSTANT assumes a quoted constant name.

Combining them leaves us with a stringified quotes constant name.
That is, the quotes become part of the name.

Remove use of HHVM_RC_INT() macro and register the constants directly.

Reviewed By: fredemmott, JoelMarcey

Differential Revision: D2757999

fb-gh-sync-id: a51657653c943625404b171ec7a3c054ae33b316