Extend documentation for invariant and friends.
Summary:
We didn't have full documentation for `invariant` in systemlib; so
I combined what was in our `hhi` with what we had there to create a
more full documentation for `invariant`, along with other friends.

The `hhi` file will now basically be a pointer to the docs which will
now contain the info from systemlib.

Reviewed By: int3

Differential Revision: D2757646

fb-gh-sync-id: 3fa0e004f87a24b3382c60ec8fb77b8c53cc8836