Try to not break tracelets in the middle of new minstrs
Summary:
Due to the number of gen-time optimizations that still exist in
irgen-minstr, we generate suboptimal code if a region has control flow diamonds
or breaks in the middle of a new minstr. Fixing those is on the long-term TODO
list, but this closes part of the gap for now. Once we get old and new minstrs
generating the same (or at least equally performant) code, we can start getting
fancy.

Reviewed By: ottoni

Differential Revision: D2760305

fb-gh-sync-id: d9723f6372a2080777edac47ee810e6b6f186f07