Pass -DNDEBUG to RelWithDebInfo's CFLAGS
"Rel" in "RelWithDebInfo" stands for "release", so this clearly this
isn't a debug build. Plus, there is currently no way to build a regular,
non-DEBUG build but with debugging symbols and RelWithDebInfo should be
just that.

Note that there is still another difference between Release and
RelWithDebInfo: -O3 vs. -O2, respectively. This patch does not change
that, as presumably an -O2 build is probably a better tradeoff to those
who are interested in a production build that allows debugging in
realtime.

We've been carrying this patch in the Debian packages for a while, as in
Debian we're interested in a release build but also shipping debugging
symbols, albeit in a separate package (hhvm-dbg).