Fix buggy FTRACE in refcount-opts.cpp
Summary:
sink_incs in refcount-ops.cpp has a FTRACE statement with three
placeholders in the format string, but with only two parameters provided. This
causes "interesting" behavior, such as generating different IR depending on
whether tracing is turned on or not. Looks like a copy-pasto from the previous
FTRACE, adjust the format string to be correct.

Reviewed By: edwinsmith

Differential Revision: D2769399

fb-gh-sync-id: 1e9820aad3e3de4358f4f1b5d47a2571382cf867