Pass -DNDEBUG to RelWithDebInfo's CFLAGS
Summary:
"Rel" in "RelWithDebInfo" stands for "release", so this clearly this
isn't a debug build. Plus, there is currently no way to build a regular,
non-DEBUG build but with debugging symbols and RelWithDebInfo should be
just that.

Note that there is still another difference between Release and
RelWithDebInfo: -O3 vs. -O2, respectively. This patch does not change
that, as presumably an -O2 build is probably a better tradeoff to those
who are interested in a production build that allows debugging in
realtime.

We've been carrying this patch in the Debian packages for a while, as in
Debian we're interested in a release build but also shipping debugging
symbols, albeit in a separate package (hhvm-dbg).
Closes https://github.com/facebook/hhvm/pull/6673

Reviewed By: svcscm

Differential Revision: D2770109

Pulled By: fredemmott

fb-gh-sync-id: 0125b7c4075f65872f90c5ded288e007147f6063