Don't lookup the thing we're looking at
Summary:
This happens once per invocation of hhvm - so I don't expect it to
help perf. It might help warmup time, but again, I wouldn't expect it
to be measurable. But this was so bizarre I spent several minutes
chasing the code just to be absolutely sure it was really doing what
it looks like its doing, so hopefully fixing it will save some future
engineering time.

Reviewed By: swtaarrs, koda, jasone

Differential Revision: D2774470

fb-gh-sync-id: 9063c55cbc7e5b960814648b358206ec0ebd52f5