Move NamingGlobal.GEnv to shared hash table
Summary: The NamingGlobal.GEnv grows with the size of the codebase since it contains naming information for every global entity (classes, constants, functions, typedefs). Passing this information to workers via `GlobalStorage` is wasteful since a worker often only needs a very small amount of data. The solution is to move this information from a map stored in the `ServerEnv` to values stored in the shared hash table.

Reviewed By: int3

Differential Revision: D2778724

fb-gh-sync-id: 566623c738107dfef725fd437760b1ad187bfa1a