enum APCKind instead of DataType
Summary:
Replace the hard-to-understand combination of DataType and flags
with a single enum. The APCHandles that are embedded in APCTypedValue
still need to set DataType to something so we can pun an APCTypedValue*
into a TypedValue*, but the other representations don't; changed them
to kInvalidDataType.

Refactored code to use APCHandle::kind() instead of type(). This lessens
the need for some flags; they'll be removed later.

Reviewed By: koda

Differential Revision: D2740566

fb-gh-sync-id: 2da594a0b827587d8d1af9536713d32798e0b28a