Get rid of a whole bunch of strcpy() invocations
Summary:
These uses are all actually fine, in theory, but we're
using a fixed buffer allocation in the driver and are
populating it (in a few cases) from APIs we don't control.

Add two new PDO driver APIs:
setPDOErrorNone() which is specialized for clearing the error.
setPDOError() which will throw a notice if the input is too large.

Reviewed By: JoelMarcey

Differential Revision: D2787374

fb-gh-sync-id: 83efc75b170925bd3929d6136ddb12cc03b2ba22