Support ciphers in openssl_seal()/openssl_open() that require an IV
Summary: There has been a longstanding bug in both HHVM and PHP where any cipher that requires an IV will fail when used in openssl_seal() and openssl_open() (https://bugs.php.net/bug.php?id=70438). In PHP 7 this is fixed by adding a reference parameter to openssl_seal() that will return the generated IV. The semantics of this are a bit confusing, but we should maintain compatability.

Reviewed By: sgolemon

Differential Revision: D2788754

fb-gh-sync-id: c866fa4345370297f10445689e0425acb45a00fd