Add common subexpression elimination pass
This diff adds a local common subexpression elimination vasm pass. It has basic
block scope and works as described below:

* searches for binary expressions inside a BB and saves them into a list
* an instruction is considered to be a binary expression if it uses two operands
and defines at least one
* if an expression is already in the saved expression list, it's replaced by a
"nop" instruction
* all uses of registers which are defined by a redundant expression are replaced
with registers defined by the matching expression (i.e. the expression found in
the list)
* if a register used by a saved expression is redefined by another instruction,
the binary expression is removed from the list. This can only happen for
physical registers, which are not in SSA form
* SF registers are also considered non-SSA, since vasm-xls does not support
overlapping liveness intervals for them.