Remove call to external commands and various fixes on Windows
Summary:
This PR include fixes as:

    - replacing all the external commands like `diff` and `find`;
    - replacing the call to `Unix.system` by `Sys.command` and `Unix.create_process`.

Note: for the `diff` command, there isn't -u/--label option on Windows, we just do the diff without this extra arguments.

PS: This work was done in order to make version 0.19.1 work on
Windows. We couldn't rebase it as it depends on previous PR that have
not been merged in master yet.
Closes https://github.com/facebook/flow/pull/1182

Reviewed By: jeffmo

Differential Revision: D2790723

Pulled By: gabelevi

fb-gh-sync-id: d3e6cb9d6aa9e8fbc6244639d15cd59c69d5f2f2