Correctly set register arguments on lookupDestructor call
The 'this' pointer can be overwritten by the unique stub
emitFreeLocalsHelpers if registers need to be allocated.

There was no issue on X64 as no additional registers are needed on that
context at the moment but it was a dorment bug. If any other
architecture port is done by using X64 as base, this change is also
helpful.