Fix cronolog
Summary:
Calling lstat on a null pointer is undefined behavior, so gcc 4.9
deduces that prevlinkname cannot be null, and always takes the rename
path, rather than the unlink path.

Pretty cool optimization, but it would work better if lstat actually
crashed when given a null pointer.

rename is apparently a no-op when the target name is a null pointer,
so we then fail to create a symlink because linkname already exists.

Reviewed By: ottoni

Differential Revision: D2814055

fb-gh-sync-id: 5fe33435dbfb60a3cdb1169331157b96b8b80134