Fix bug in C++ ini test
Summary:
When documenting our ini settings, I came across

`hhvm.server.high_priority_end_points`

When I grepped for `HighPriorityEndPoints` I found this test and
then my eye was grabbed by the fact that the actual name in
`TestCollectionIni()` was *wrong*. It is `end_points`, not
`endpoints`.

So why was this test passing? Well, because when I wrote it I forgot
to clear out that runtime option from the HDF run we did before. So
I did that too here. Made the test case just a tad different from the
HDF run and all is right in the world again.

I also reordered the HDF test to be closer to its cousin as well.

Reviewed By: markw65

Differential Revision: D2814362

fb-gh-sync-id: db12814627733c5594385c190906408996963a9f