Correctly set register arguments on lookupDestructor call
Summary:
The 'this' pointer can be overwritten by the unique stub
emitFreeLocalsHelpers if registers need to be allocated.

There was no issue on X64 as no additional registers are needed on that
context at the moment but it was a dorment bug. If any other
architecture port is done by using X64 as base, this change is also
helpful.
Closes https://github.com/facebook/hhvm/pull/6717

Reviewed By: mxw, paulbiss

Differential Revision: D2816024

fb-gh-sync-id: 723f494a7e93e4a2b48a04b4df2296d263a09e22