Do not attempt to autoload already loaded classes
Summary:
`TypeStructure::getAlias` is used to get the type structure associated with a
type alias. This would invoke `Unit::loadTypealias` and attempt to load the
type alias. `TypeStructure::getAlias` could be called on either an alias or a
class, but  `Unit::loadTypealias` does not check to see if the name maps to an
already loaded class before invoking the autoloader.

To fix this I check if the name is an already loaded class using
`Unit::lookupClass`, if so then there is no associated type structure so we
return an empty array.

Reviewed By: swtaarrs

Differential Revision: D2813757

fb-gh-sync-id: aafb0cf1c2d16c7ec107930af22a95a97c610506