APC snapshot priming
Summary:
Add snapshot support for APC priming, to load faster and with less memory than current shared library format.

The new format is auto-detected when passed to the existing PrimeLibrary flag (it can even keep the .so extension).

To "upgrade" an existing .so to the new format, there is a new flag that just writes the snapshot to a file and then exits hhvm (without starting any servers).

Reviewed By: binliu19

Differential Revision: D2826183

fb-gh-sync-id: c083192d94c117fc3a7234988b23f8e9208814c6