Fix typo in null-coalesce implementation
Summary:
The implemention of in() in hhbbc/interp.cpp for the CGetNQuiet
bytecode should be identical to that of CGetN. The intention was that we could
re-use the same implemention by just calling the CGetN implementation. Since the
CGetN implementation did not actually refer to anything in the opcode, we could
obtain the proper overloading behavior by just casting a null to a
CGetN. Unfortunately, a typo meant that we casted to CGetNQuiet instead, meaning
we just recursively call ourself infinitely.

Rather than just fix the typo, avoid the hackery of casting a nullptr by
factoring out the CGetN implementation into a separate function and having both
CGetN and CGetNQuiet call that. It also turns out that the implementations
shouldn't be quite identical, they need a template parameter to determine which
CGetL equivalent they should be optimized to.

Also add a test case to actually exercise these different "quiet" versions of
CGet.

Reviewed By: markw65

Differential Revision: D2850127

fb-gh-sync-id: 68d793313f1bfbaada5a723d078b81a98f10ee03