Make ReflectionParameter::getDefaultValueConstantName compatible with
Summary:
`ReflectionParameter::getDefaultValueConstantName()` was being overloaded to do
things it was not supposed to do.

http://php.net/manual/en/reflectionparameter.getdefaultvalueconstantname.php

Currently it just returns the raw text for a default value. It should only
return the text if the default value is a *constant*.

If we have:

```
define("MY_CONST", 4);
function foo($a = MY_CONST, $b = MY_CONST * MY_CONST, $c = array(MY_CONST));
```

`ReflectionParameter::getDefaultValueConstantName()` should only return a value
for `$a`. Not the others, like it is now.

So our custom `ReflectionParameter::getDefaultValueText()` should not have been deprecated.

Reviewed By: fredemmott

Differential Revision: D2831292

fb-gh-sync-id: f6a5534d74687d10f3ca9cd955f7601630fd988f