Match native and stublogue{} frame records
Summary:
I'm pretty sure the only reason these weren't in line to begin with was
that x64 and ARM have the same frame record layout, and I didn't want to
make a separate vasm instruction for the transition between stublogue{}
and phplogue{} needed for fcallArrayHelper.

PPC64 has a different frame record, so let's add a stubtophp{}
instruction and change the stublogue{} ABI to be arch-dependent w.r.t.
its stack layout.

Reviewed By: markw65

Differential Revision: D2851925

fb-gh-sync-id: b1fdcf8b35889cd8411341bd2a92a6d07b49af8b