Fix lookupUnitRepoAuth
Summary:
If anything threw while it was loading the unit, the unit would be
forever marked as not existing.

Originally, nothing should have thrown, but we added surprise checks
to the unserialize code, so timeouts, and ooms could end up throwing.

This only affects RepoAuthoritative mode; in sandbox mode, we'll just
try again if we find an entry with a null unit.

Reviewed By: swtaarrs

Differential Revision: D2867957

fb-gh-sync-id: 51ace1ae6a4134c5df461983e428535227cc2732