Use lea{} instead of addqi{}/subqi{} for rsp()
Summary:
This makes things a bit nicer for non-x64 land, where there are
sometimes further restrictions on the native stack pointer.

Two of these changes are in very slow (intercept-related) paths in
unique stubs, which is surely not perf-sensitive.

The other two are for lowering vcall{}.  Interestingly, we already use
lea{} for pushing vcall{}'s args, just not for popping them.  Probably
doing both won't be the end of the world, and if perflab is happy, I
think this is Fine (TM).

Reviewed By: markw65

Differential Revision: D2869189

fb-gh-sync-id: 649ea58f11e4e34556456b4c0c7ccf467a3e3f83