Invoke autoloader when constructing ReflectionTypeAlias/type_structure
Summary:
For whatever reason when we implemented refelection support for type aliases we
did not attempt to autoload the type alias, but instead looked it up in the
`NamedEntity` map. This means makes the `ReflectionTypeAlias` class also
useless because if the referenced type alias was not autoloaded it would fail
to construct a reflection object. The same logic was copy over to
`type_structure` when we changed the function to support fetching the type
structure for a type alias as well.

This diff fixes this bug by using `Unit::loadTypeAlias` instead of directly
using the `NamedEntity` table.

Reviewed By: paulbiss

Differential Revision: D2871069

fb-gh-sync-id: 791d5068d69612f846bbd7c54502ea91437c6958