Add internal_error as a replacement for `assert false`
Summary:
I've been wanting to do this for a while. Aside from the obvious benefit of
avoiding typechecker crashes when "impossible" things happen, `internal_error`
also can serve as a convenient way for getting information out of the worker
processes (just stringify the info and put it in the error message). I think it
will be useful for my refactoring of decl.

I only removed a handful of `assert false`s for now because I'm lazy.

Reviewed By: dabek

Differential Revision: D2874362

fb-gh-sync-id: b30ff9843e0ea19c367f40a5899c799bb878f186