Fix crash on aborted rpc request
Summary:
rpc threads are left with g_context != null between requests, so
generally there's no need to set it up in abortRequest. But if the
first request to a thread is aborted, we do need to.

Also, we mark a thread for reset after a fatal, so we should probably
do the same for an aborted request.

Reviewed By: mxw

Differential Revision: D2874659

fb-gh-sync-id: faeb62db80f2420c1a0ad64675738ae2f0935afe