Refactor .eh_frame CFI record emission
Summary:
Refactor the implementation of register_unwind_region() so that we can
do similar work at vasm emit-time.

There's some similar instrumentation in debug/dwarf.{h,cpp}, and most of
it is dead.  I've deleted those bits, but am saving rewriting the rest
to use EHFrameWriter for a separate diff.

This ever-so-slightly changes the existing TC CIE, by adding an address
encoding for the FDE's pointers (augmentation "R").  The default appears
to be DW_EH_PE_absptr (which has the value 0x0), but this behavior isn't
documented anywhere, so it feels better not to rely on it.

Reviewed By: swtaarrs

Differential Revision: D2880982

fb-gh-sync-id: a77111b2f18336147596b9f1b19a7c6e56f8933b