Fixing async stack trace when await multiple dependencies.
Summary:
Currently, AwaitAllWaitHandle only establishes dependencies on the first unfinished child, even if we have multiple children still running. This is beneficial for efficiency as notifications are only needed when that specific dependent child finishes.

However, the downside of this approach is, for all other children except the first unfinished one, the parent function is missing from their async stack (as the dependency not built). This could cause troubles for e.g. catching regression or debugging, as the responsible function might be missing from stack trace.

This diff is to build proper async dependencies for async functions, so that we have correct stack traces. Also, try to minimize the impact from doing so.

Reviewed By: jano

Differential Revision: D2763858

fb-gh-sync-id: 9a4573021604123b5ba8fc71df6a51fae8fb43f3