Fix loadqp on ppc64 for pointer > 4GB address space
Summary:
The RIP address might have a displacement that is above the 4GB address space
(32 bits), the baseless cast from MemoryRef to Vptr would incorrectly cast it
to a lower precision type (safe_cast would be a nice candidate there).

This solution places the displacement parameter directly on a register for a
direct load.
Closes https://github.com/facebook/hhvm/pull/6772

Reviewed By: mxw

Differential Revision: D2846473

fb-gh-sync-id: 961c2df1d0ba7f82830a3c0cff6bc0279f3bbacd