Use ftello64 for explicit 64-bit offsets
Summary:
`ftello` returns an `off_t`, which, under MSVC at least, is typedef'd to `long`. For MSVC that means the return value would be 32-bit, but the use of the function is expecting it to be 64-bit, so use `ftello64` which returns an `off64_t`, which is explicitly 64-bit.

As a random note, Windows doesn't actually implement `ftello` or `ftello64`, but I want to use the correct API when implementing the file portability stub.

Reviewed By: paulbiss

Differential Revision: D2860640

fb-gh-sync-id: 3d6ed942da45782ea3aea0649d5ea8bdf30f64f8