Fix ThreadLocalSingleton destruction hook
Summary:
The fast-TLS version of ThreadLocalSingleton was completely ignoring the destruction hook on thread exit. In particular, ~MemoryManager was never called. This was preventing the addition of automatic leak checks.

Fix this by making ThreadLocalSingleton use the ThreadLocalManager, ensuring its exit hook is called just like other thread locals. When the heap scanner treats the TLMs list as roots, it skips the MemoryManager.

Also add a first unit test for ThreadLocalSingleton.

Reviewed By: markw65

Differential Revision: D2883813

fb-gh-sync-id: bcfe82c33d7ab005e2f05e3d9168f27087e42950