Avoid throwing if droot is not set
Summary:
droot is the class / function / typedef declaration that we are currently
checking. We look it up whenever we need to add a dependency, e.g. to say that
class A depends on class B and must be rechecked whenever the latter changes.

We've seen some cases of hh_server crashing due to code trying to add
dependencies when droot was not set. It turns out that there are legitimate
cases when it isn't set; for example, calling sub_type / unify can result
in the dependency adding code being called, but we directly call sub_type while
running the Typing_suggest_service, i.e. when we are not checking any
particular class / function / typedef declaration.

To avoid bugs where droot is accidentally left unset, we now require it to be set
(or explicitly set to `None`) when the typing env is created. `set_root` is removed
entirely -- there is no need to change the root after the typing env has been
created.

Reviewed By: dlreeves

Differential Revision: D2885310

fb-gh-sync-id: af68486582950972e2bd03f0167aec30f529ca6a