Add testing scaffold for dependencies
Summary:
It's been a fairly opaque part of our stack. The lack of a solid way to test
changes to it has led to a number of bugs that were only discovered after
deployments. Will also be useful as a sanity check during as I refactor decl
mode.

I already see something suspicious in the test output: there are dependencies
on both the namespace-slash-prefixed and non-prefixed forms of the class.

I made `--no-builtins` a bool setting instead of a mode because there can only
be one active mode at any particular point in time. We want to disable builtins
for our dependency tests to avoid seeing a massive dump of builtins dependencies
in the test output.

Reviewed By: dlreeves

Differential Revision: D2884656

fb-gh-sync-id: 3c6f5b8458e1c0e67d681665419695afbcf9ceaf