General cleanup of ext_pdo
Summary:
The biggest thing this does is make it so that ext_pdo_sqlite and ext_pdo_mysql aren't used in ext_pdo. This also eliminates the need to declare the individual PDO drivers in pdo_driver.cpp, and instead allows them to be declared in the respective extension, ext_pdo_sqlite / ext_pdo_mysql.

This also eliminates all the global constants in C++ that are not actually needed, and removes the global StaticStrings that were only used once.

Reviewed By: paulbiss

Differential Revision: D2870465

Pulled By: Orvid

fb-gh-sync-id: d89b7e7d27a989c74115c1793bbe810fb41f32b8