Reverted commit D2870465
Summary:
The biggest thing this does is make it so that ext_pdo_sqlite and ext_pdo_mysql aren't used in ext_pdo. This also eliminates the need to declare the individual PDO drivers in pdo_driver.cpp, and instead allows them to be declared in the respective extension, ext_pdo_sqlite / ext_pdo_mysql.

This also eliminates all the global constants in C++ that are not actually needed, and removes the global StaticStrings that were only used once.

Reviewed By: paulbiss

Differential Revision: D2870465

Pulled By: swtaarrs

fb-gh-sync-id: 2524eb5b8b9fd94a6fbe17a0925981362a30eab0