rename Utils.String
Summary: shadowing `String` in a module that is so commonly `open`ed seems to sometimes cause incompatibility in modules that don't `open Utils`. or maybe string literals (sugar for a `String.t` i think?) created in a file that doesn't `open Utils` are incompatible with the `String.t`'s from files that do? I'm not sure exactly but I kept running into errors about `string` not being compatible with `Utils.SSet.t`. Either way, it seems better to avoid shadowing and `include` if we don't need it.

Reviewed By: int3

Differential Revision: D2915934

fb-gh-sync-id: 6941570b109d233a5844ad4f5b4ad38f7a78ecde
shipit-source-id: 6941570b109d233a5844ad4f5b4ad38f7a78ecde