move maps and sets out of Utils
Summary:
`Utils` is a random assortment of stuff, and by far the most common reason it is required is for `SSet` and `SMap`. this diff moves `MyMap`, `HashSet`, `SMap`, `IMap`, `SSet`, `ISet`, and `CSet` into their own directory as modules, so that you don't have to `open Utils` to get them.

all of the interesting changes (if any) are in `hphp/hack/src/utils/collections/`, the rest is just codemodding away the `Utils.` or `open Utils` that are no longer necessary, and changing `MyMap` (a functor) and `MapSig` to `MyMap.Make` and `MyMap.S` (just like the built-in `Map`).

Reviewed By: int3

Differential Revision: D2916292

fb-gh-sync-id: 6f55489d2fa3f7dcb15c01f89313063e8947197f
shipit-source-id: 6f55489d2fa3f7dcb15c01f89313063e8947197f