Added typehints and refined mysqli implementation
Some of the typehints in the mysqli implementation were not real,
according to the C++ implementation and the official PHP docs, so
those incorrect typehints are now changed. I added typehints to
the typechecker definition for mysqli and adjusted functions and
methods that used references before; I've also removed some methods
and functions that seem not to be implemented in Hack.