Moving threshold callback to event-hooks
Summary:It wasn't safe when it was called from check_surprise_flags.
It was previously called from locations that do not support re-entrancy and since the callback does not unwind the stack re-entrancy is possible and in rare cases crashes HHVM.
This was mostly common in cases of JSON Ser/Deserialization.

Reviewed By: markw65

Differential Revision: D2906277

fb-gh-sync-id: d943aa08cbcc6b27887aec34d6929487c7b7975c
shipit-source-id: d943aa08cbcc6b27887aec34d6929487c7b7975c