Create processes using Daemon.fork instead of Daemon.spawn (temporarily)
Summary:The plan for accessing shared memory from IDE process was to move SharedMem.init line from ServerMain to HhServeverMonitor, so it would be initialized before those processes are thus shared between them.

It turns out that this doesn't work with Daemon.spawn - there's a work in progress to make it work, but it's been in review for more than a month and I don't want to be blocked on it. After #9556242 is done I can revert this diff and go back to approach from D2879026.

Reviewed By: int3

Differential Revision: D2890078

fb-gh-sync-id: c281f5e292fd1b5f55a0113857230ad34ddb54c2
shipit-source-id: c281f5e292fd1b5f55a0113857230ad34ddb54c2