Fixing async stack trace when await multiple dependencies.
Summary:*This is the re-commit of D2763858 -- everything is identical except the stack trace test is updated in cycle_debug_backtrace.php.expectf.

Currently, AwaitAllWaitHandle only establishes dependencies on the first unfinished child, even if we have multiple children still running. This is beneficial for efficiency as notifications are only needed when that specific dependent child finishes.

However, the downside of this approach is, for all other children except the first unfinished one, the parent function is missing from their async stack (as the dependency not built). This could cause troubles for e.g. catching regression or debugging, as the responsible function might be missing from stack trace.

This diff is to build proper async dependencies for async functions, so that we have correct stack traces. Also, try to minimize the impact from doing so.

Reviewed By: jano

Differential Revision: D2936352

fb-gh-sync-id: da7c2b1bec2ed39227e4e4fc08e580f1022fa5da
shipit-source-id: da7c2b1bec2ed39227e4e4fc08e580f1022fa5da