Sweep using PtrMap instead of a second heap iteration.
Summary:Heap iteration is the slowest way to visit everything in the heap.
Using PtrMap is faster because the pointers themselves are stored
in a dense array, and we only visit the non-free blocks we added to
the map during init.

During sweep, since we're iterating over PtrMap instead of the heap
itself, it's safe to free children objects during the final iteration
instead of batching them up and freeing them after the heap iteration
is over.

Add Quarantine option, use it instead of EagerGC to enable quarantine.

Also add resetEagerGC(), which I forgot to add earler; this clears
the eager-mode pointer filter between requests.

Add NEVER_INLINE to Marker::init(), trace(), sweep() so the major
gc phases are visible in perf.

Reviewed By: ricklavoie

Differential Revision: D2932230

fb-gh-sync-id: 221d8d1be22916cacd4f6ced83e376a58108dda4
shipit-source-id: 221d8d1be22916cacd4f6ced83e376a58108dda4