Add an option to use local memory instead of shared one, use it in autocomplete to make it safe to run in parallel
Summary:Most of our code treats shared memory as global variable, which is written and read freely in many functions, including those that operate on local IDE buffers. IDE code deals with it by doing "oldify and revivie" dance - identifies the definitions that might be changed, moves them to a different place, and afterwards brings them back (or not, I'm pretty sure we have some bugs with that, see #9437856).

The disadvantage of this, is that it incurs an "unnecessary" write to global memory, and those are not always safe to do in parallel. In this diff I introduce a workaround to this, by intercepting all the writes into process local hashtable and never modifying the global one.

Reviewed By: int3

Differential Revision: D2894948

fb-gh-sync-id: 03ef7f7995d323bd1e752a58a38a6e59ce50edd1
shipit-source-id: 03ef7f7995d323bd1e752a58a38a6e59ce50edd1