Blacklist xdebug test in phpunit
Summary:Normally the testMessageWithXdebugScreamOn test is skipped as xdebug is
disabled.  Sandcastle is somehow running this test though...  I poked around on
a sandcastle machine and I don't see why it would enable xdebug,  and I can't
replicate it locally on my devserver, so just blacklist the test file for now.

Reviewed By: mxw

Differential Revision: D2941641

fb-gh-sync-id: 124f168e77aba927aa7f0622ef63d9345a38328c
shipit-source-id: 124f168e77aba927aa7f0622ef63d9345a38328c