Fix Func AttrHot initialization
Summary:It's possible for hhvm to warm up without ever seeing a request where
(profileOn && !p) is true (in profileRequestStart()). When this happened,
nothing was calling setHotFuncAttr(), so no Funcs were marked hot and ahot was
sitting there empty, twiddling its thumbs. Rather than waiting for a condition
that might never happen, just do a quick check before we need to determine if a
Func is hot. This is exactly what we currently do with InstanceBits::init().

Reviewed By: markw65

Differential Revision: D2942572

fb-gh-sync-id: 79140e553d70f78785d2ed558847b54efd9b2e65
shipit-source-id: 79140e553d70f78785d2ed558847b54efd9b2e65