Cleanup the Clang and GCC sections of HPHPCompiler.cmake
Summary:A 500 column line with all of the options for GCC is not very easy to read, and also means quite a few options have to be duplicated.
This refactors things to work in a way that is much closer to the MSVC section, although the MSVC section is far more complex.

Reviewed By: JoelMarcey

Differential Revision: D2927425

fb-gh-sync-id: 228fe728cde6d0873d82b342bb80a79c0be043d4
shipit-source-id: 228fe728cde6d0873d82b342bb80a79c0be043d4