Split make_param_ty into two separate functions for decl vs locl
Summary:As part of lazy decl, I'm trying to make Typing depend on Typing_decl instead
of the other way around.

One of the functions causes Typing_decl to call into Typing is make_param_ty,
and it seems like a fairly frankensteinian function -- it takes a bunch of
parameters in order to cover both the decl and locl phases. Splitting the
function into two makes it easier to read and easier to split between the
Typing_decl and Typing modules.

One minor functionality change is in the way we handle `= null` in the param list. If we see this

```
function f(int $x = null) {}
```

We raise an error (since $x is typehinted as non-null). However, due to
backwards compat reasons, we would previously continue to typecheck the body
of `f` with the assumption that `$x` is `?int`. In the interests of
simplicity, I've done away with that logic; we will now assume that `$x` is
just `int` in the body of `f`.

In any case, this feature seems pretty old and the backwards compatibility
may no longer be necessary, so I think it's alright to change it.

Reviewed By: dlreeves

Differential Revision: D2947924

fb-gh-sync-id: 743e4a495454604ce8e3fbb5f0dafcae3f642e63
shipit-source-id: 743e4a495454604ce8e3fbb5f0dafcae3f642e63