Reject large content-length early, in proxygen-transport
Summary:Reject POST requests with a large content-length early,
just after we receive headers, but before we accept all the
data from the client. In order to use the proper value for
the post size limit, factor virtual host resolution so we
can call it from proxygen-transport, without changing the way
state is managed in http-protocol.cpp.

This only affects proxygen-transport, not fastcgi or any other
http transport.

Reviewed By: markw65, afrind

Differential Revision: D2925473

fb-gh-sync-id: 932f815a62bef82f903df29eb3d19d6f069f1533
shipit-source-id: 932f815a62bef82f903df29eb3d19d6f069f1533