Do not ignore declared variance of generics with constraints
Summary:Let's take a look at this simplified example

```
class Foo<+T as int> {
  public function bar(T): void {}
}
```

Previously we would allow this code to type check because we ignored the
variance specified on `T` since it has a constraint. Reading the comments the
reason was that the constraint does not have the same variance as the generic
itself, but is contravariant for `as` constraints, yet covariant for `super`.
While this is true we should not ignore the specified variance of the generic
like we currently are doing, instead we should include the variance of the
generic, but add additional variance constraints for the `as` or `super`
constraint.

Reviewed By: int3

Differential Revision: D2922621

fb-gh-sync-id: ad324b29cfc1259c7c7da1e3fb4d36d5775ffeec
shipit-source-id: ad324b29cfc1259c7c7da1e3fb4d36d5775ffeec