Add explicit indirect return arguments.
There are stark differences between the x86_64 and AArch64 ABIs
with regards to indirect return arguments. This work is similar
to the work done for the interpreter when calling native functions.
In this case though, the ArgGroups made the port a little more
difficult. This change expands the ArgGroups to handle indirect
return arguments explicitly when the target supports them.
Otherwise, the behavior for targets without dedicated regsiters for
indirect return arguments is to pass it as the first general
purpose argument.

Fixes (AArch64):
   test/slow/php7_backported/ext_openssl/openssl_encrypt_ccm.php
   test/slow/php7_backported/ext_openssl/openssl_encrypt_gcm.php