Always partition by text area in layoutBlocks()
Summary:In practice, this shouldn't make much of a difference, but it's a useful
restriction when we're generating CFI, so that we don't have have multiple live
DWARF buffers at once during vasm-emit.

What we actually care about is partitioning by /actual/ code region, not just
the AreaIndex (since, e.g., Main and Cold might be aliased), so add some
asserts and code and stuff.

Reviewed By: ottoni

Differential Revision: D2952273

fb-gh-sync-id: f0e978317f7dbfa52f2caa97c5174762f5b81b3c
shipit-source-id: f0e978317f7dbfa52f2caa97c5174762f5b81b3c