Treat connect_to_monitor timeout as ServerMissing (and attempt to start it)
Summary:We use this call to tell whether server is running, assuming that it will immediately succeed or error. Recently it started timing out. In all of the cases I've investigated all servers are dead, but the kernel claims that something is listening on this socket. I don't know why it suddenly stopped cleaning it up immediately. One of the suspects is {D2890078} where fork related fd duplication might cause it to happen, another one is some kernel change.

Whatever it is, restarting the server should help - it uses SO_REUSEADDR option which means that it should succeed. If the timeout was caused by some other issue, and server was already running, attempt to start it will fail due to lock file.

Reviewed By: alexchow

Differential Revision: D2968804

fb-gh-sync-id: 919c9c8346a6f1a3d5249731711a97bf4aebb928
shipit-source-id: 919c9c8346a6f1a3d5249731711a97bf4aebb928