Nuke MCGenerator's public API
Summary:After removing all the single-callsite function members, it turns out
that MCGenerator is a class that...
...
...
...
generates machine code.  Oh, and allocates stubs in frozen that
everybody implicitly expects to be always svcreq::stub_size() in size.

Most of this diff is deleting unused code or moving single-callsite code
inline with the callsite.  The more notable include:
- Removed "get" from MCGenerator accessors.
- Renamed s_perfCounters to g_perf_counters, and moved relevant code to
  a separate header and source.
- Moved TC info stringifiers into a separate header and source.
- Moved stack overflow handlers into a separate header and source.
- Moved function-entry-helpers-x64.cpp to unique-stubs.cpp.  I also
  don't see what's terribly x64-specific about them, which is
  fortunate.  (Or I'm wrong, which is unfortunate.)

Oh also: lots of documentation.

Reviewed By: swtaarrs

Differential Revision: D2976121

fb-gh-sync-id: 71c324bc926c8ed53d722246437a0a9f0492e86d
shipit-source-id: 71c324bc926c8ed53d722246437a0a9f0492e86d