Allow type constants defined in abstract final classes to be referenced
Summary:Currently we have a bug where you cannot reference a type constant that is
declared within an abstract final class (even though the type constant isn't
abstract itself). For examples

```
abstract final class Foo {
  const type T = int;
}
```

We are emitting an error when `Foo::T` is used.

I updated our visitor we use that checks instantiability to now check if the
root of a type access is instantiable, fixing this error.

Reviewed By: dabek

Differential Revision: D2665907

fb-gh-sync-id: 139ce892332f9726db6f26ba6e6452858278abc6
shipit-source-id: 139ce892332f9726db6f26ba6e6452858278abc6