Fix maps with void* keys
Summary:There's a few uses of maps with void* keys, where the void* are actually
pointers to countable objects (ArrayData, ObjectData, etc). This causes issues
with the type-scanner machinery, as it does not know that the void* point to
countable req-heap allocations. Improve the type safety of these maps by instead
using a "PtrWrapper" class as the key. These classes are just a union with all
the possible different countable object pointers with an implicit conversion
from them. This generates the same code as using a bare void*, but retains type
information.

The use of void* for keys in php_yaml_init.h can just be replaced by the use of
HashTable*, since that's all that's ever actually passed in.

Reviewed By: alexmalyshev

Differential Revision: D2982423

fb-gh-sync-id: 40cfc739b44ca22260d9f940f506216e294351a5
shipit-source-id: 40cfc739b44ca22260d9f940f506216e294351a5