Fix heap overflow(s) in iptcembed
Summary:This code was truncating the size of the file to 32 bits. It also
wasn't checking to make sure the fstat() call actually succeeded, which it
won't for certain types of stream wrappers. It's still not perfect, but a
proper fix will require a much more disruptive change.

I also found an incorrect hardcoded buffer size thanks to ASAN that's been
fixed as well. It looks like the original author didn't realize that `\x008B`
is a single character in a string literal (this has been fixed in php-src since
we imported the extension). I also took out the final `\0` from the string so
we can just use `sizeof(psheader)`, relying on the implicit null terminator.

Reviewed By: markw65

Differential Revision: D2880195

fb-gh-sync-id: 3da5f1eaaa511298c11b0a085a4582849cb025fb
shipit-source-id: 3da5f1eaaa511298c11b0a085a4582849cb025fb